+++ packages/public-api/src/systemApiRouter.ts
@@ -1,4 +1,4 @@
-import { SystemAPI } from '@gradejs-public/shared';
+import { systemApi } from '@gradejs-public/shared';
 import { Router } from 'express';
 import { syncWebPageScanResult } from './website/service';
 
@@ -6,7 +6,7 @@ const systemApiRouter = Router();
 
 systemApiRouter.post('/scan', async (req, res, next) => {
   try {
-    const scanReport = SystemAPI.apiScanReportSchema.parse(req.body);
+    const scanReport = systemApi.apiScanReportSchema.parse(req.body);
     await syncWebPageScanResult(scanReport);
   } catch (e) {
     next(e);
